# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fleet API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/approvals": {
      "get": {
        "tags": [
          "approvals"
        ],
        "summary": "List Approvals",
        "operationId": "list_approvals_v1_approvals_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "pending",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApprovalOut"
                  },
                  "title": "Response List Approvals V1 Approvals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/approvals/{approval_id}/decide": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Decide Approval",
        "operationId": "decide_approval_v1_approvals__approval_id__decide_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Approval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/service/email-send": {
      "post": {
        "tags": [
          "service"
        ],
        "summary": "Email Send",
        "description": "Queue an email for human approval. **Never sends.**\n\n`email.send` is write:external (TRD \u00a79 names customer email as the\ncanonical always-approved case), so the only thing this endpoint can do is\ncreate the queue entry; the message goes on the wire in the approvals\nrouter's resume handler, after a person approves it.",
        "operationId": "email_send_v1_service_email_send_post",
        "parameters": [
          {
            "name": "x-fleet-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fleet-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailSendIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailSendOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "agent_id": {
            "type": "integer",
            "title": "Agent Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "sla_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agent_id",
          "run_id",
          "action",
          "payload",
          "status",
          "decided_by",
          "decided_at",
          "sla_at",
          "created_at"
        ],
        "title": "ApprovalOut"
      },
...
      "DecisionIn": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "edited_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Payload"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "DecisionIn"
      },
...
      "EmailSendIn": {
        "properties": {
          "to": {
            "type": "string",
            "title": "To"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "to",
          "subject",
          "body"
        ],
        "title": "EmailSendIn"
      },
      "EmailSendOut": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "approval_id": {
            "type": "integer",
            "title": "Approval Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "approval_id",
          "run_id"
        ],
        "title": "EmailSendOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}